requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
click==8.1.7
colorama==0.4.4
coloredlogs==15.0.1
decorator==5.1.1
entrypoints==0.3
exceptiongroup==1.2.0
executing==0.8.2
fastapi==0.110.1
flatbuffers==24.3.25
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
humanfriendly==10.0
idna==3.6
ipython==8.0.1
jedi==0.18.1
matplotlib-inline==0.1.3
mpmath==1.3.0
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.26.4
onnxruntime==1.17.1
opencv-python-headless==4.2.0.32
packaging==24.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
pillow==10.3.0
prettytable==3.4.1
protobuf==5.26.1
pure-eval==0.2.2
pydantic==2.6.4
pydantic_core==2.16.3
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.1
stack-data==0.1.4
starlette==0.37.2
sympy==1.12
tomli==2.0.0
typing_extensions==4.11.0
uvicorn==0.29.0
watchfiles==0.21.0
wcwidth==0.2.5
websockets==12.0
